+++ debug/fileclose.mjs
@@ -17,14 +17,22 @@ let dir = '../test/fixtures'
 
 
 async function main() {
-	let names = (await fs.readdir(dir)).filter(fileFilter)
+	let names = await fs.readdir(dir)
 	let paths = names.map(name => dir + '/' + name)
-    console.log('paths', paths)
-	let promises = paths.map(path => exifr.parse(path, true))
-	await Promise.all(promises)
+	await Promise.all(paths.map(handleFile))
 	console.log('DONE')
 }
 
+async function handleFile(path) {
+	let stat = await fs.stat(path)
+	if (stat.isDirectory()) return
+	try {
+		await exifr.parse(path, true)
+	} catch(err) {
+		console.log('ERROR', path)
+		console.log(err)
+	}
+}
 
 main().catch(console.error)